requirements.txt
# Requisitos principales
openai>=1.0.0  # API de OpenAI
tiktoken>=0.5.0  # Tokenizador para OpenAI
requests>=2.31.0  # HTTP requests
python-dotenv>=1.0.0  # Carga de variables de entorno

# Reconocimiento de voz
SpeechRecognition>=3.10.0  # Reconocimiento de voz
PyAudio>=0.2.13  # Interfaz de audio (requiere portaudio en el sistema)
vosk>=0.3.45  # Reconocimiento de voz offline

# Síntesis de voz
gTTS>=2.3.2  # Google Text-to-Speech
pyttsx3>=2.90  # Síntesis de voz offline
playsound>=1.3.0  # Reproducción de sonidos
pygame>=2.5.0  # Alternativa para reproducción de audio

# Utilidades
psutil>=5.9.5  # Información del sistema
numpy>=1.24.0  # Procesamiento numérico
pydantic>=2.0.0  # Validación de datos
tenacity>=8.2.0  # Reintentos para APIs

# Interfaz de texto
rich>=13.0.0  # Formato de texto mejorado en consola
prompt_toolkit>=3.0.38  # Para interfaces de línea de comandos

# Desarrollo
pytest>=7.3.1  # Pruebas
pytest-asyncio>=0.21.0  # Pruebas asíncronas
black>=23.3.0  # Formateador de código
isort>=5.12.0  # Ordenador de imports
mypy>=1.3.0  # Verificación estática de tipos
